Refactor backend to use a better structure
